+++ Integrations/KennaV2/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## [Unreleased]
--
+ - KennaV2 moved to python and added 4 new commands:
+    - ***search assets*** 
+    - ***get vulnerabilities of a specific asset***
+    - ***add new tag to asset***
+    - ***delete new tag from asset***